fixed ns3 warnings, added pbs script and removed ns3:: from output filename

+++ ns3_igraph/ns-allinone-3.19/ns-3.19/scratch/brite-generic-example.cc
@@ -129,8 +129,8 @@ main (int argc, char *argv[])
   NodeContainer client;
   NodeContainer server;
   
-  int numClients = 0.75 * nLeafNodes;
-  int numServers = 0.25 * nLeafNodes;
+  int numClients = static_cast<int>(0.75 * nLeafNodes);
+  int numServers = static_cast<int>(0.25 * nLeafNodes);
   client.Create (numClients);
   stack.Install (client);
   server.Create (numServers);
