+++ ns3_igraph/ns-allinone-3.19/ns-3.19/src/lte/model/lte-asn1-header.cc
@@ -336,7 +336,7 @@ void Asn1Header::SerializeInteger (int n, int nmin, int nmax) const
     }
 
   // Clause 11.5.6 ITU-T X.691
-  int requiredBits = std::ceil (std::log (range) / std::log (2.0));
+  int requiredBits =static_cast<int>( std::ceil (std::log (range) / std::log (2.0)));
 
   switch (requiredBits)
     {
@@ -548,7 +548,7 @@ Buffer::Iterator Asn1Header::DeserializeInteger (int *n, int nmin, int nmax, Buf
       return bIterator;
     }
 
-  int requiredBits = std::ceil (std::log (range) / std::log (2.0));
+  int requiredBits = static_cast<int>(std::ceil (std::log (range) / std::log (2.0)));
 
   std::bitset<1> bitsRead1;
   std::bitset<2> bitsRead2;
